refactor(Query): move all the strawberry types into separate files
In the process of adding the project_groups code ran into many
instances of having to do hacky workarounds to prevent circular imports
in the schema file. So, I've moved the types to a types/ folder so they
can safely import each other with ease and the schema file can import
each GraphQL type without having to worry about circular imports.

This simplifies the schema.py to just the Query class and setting up
the graphql schema with the strawberry schema builder.
