+++ api/graphql/mutations/__init__.py
@@ -2,7 +2,6 @@ import strawberry
 from strawberry.types import Info
 
 
-from api.graphql.loaders import GraphQLContext
 from api.graphql.mutations.cohort import CohortMutations
 from api.graphql.mutations.comment import CommentMutations
 from api.graphql.mutations.project import ProjectMutations
@@ -13,9 +12,9 @@ from api.graphql.mutations.family import FamilyMutations
 from api.graphql.mutations.participant import ParticipantMutations
 from api.graphql.mutations.sequencing_group import SequencingGroupMutations
 from api.graphql.mutations.sample import SampleMutations
+from api.graphql.mutations.project_groups import ProjectGroupsMutations
 
 from api.graphql.mutations.user import UserMutations
-from models.models.project import FullWriteAccessRoles, ReadAccessRoles
 
 
 @strawberry.type
@@ -87,3 +86,9 @@ class Mutation:
     def user(self) -> UserMutations:
         """User mutations"""
         return UserMutations()
+
+    # Project Groups
+    @strawberry.field
+    def project_group(self) -> ProjectGroupsMutations:
+        """Project group mutations"""
+        return ProjectGroupsMutations()
