+++ api/graphql/mutations/assay.py
@@ -1,18 +1,17 @@
 # pylint: disable=redefined-builtin, import-outside-toplevel
 
-from typing import TYPE_CHECKING, Annotated
+from typing import Annotated
 
 import strawberry
 from strawberry.types import Info
 
 from api.graphql.loaders import GraphQLContext
+from api.graphql.types.assay import GraphQLAssay
+from api.graphql.types.comments import GraphQLComment
 from db.python.layers.assay import AssayLayer
 from db.python.layers.comment import CommentLayer
 from models.models.assay import AssayUpsert
 from models.models.comment import CommentEntityType
-
-if TYPE_CHECKING:
-    from api.graphql.schema import GraphQLComment, GraphQLAssay
 
 
 @strawberry.input  # type: ignore [misc]
@@ -36,11 +35,8 @@ class AssayMutations:
         content: str,
         id: int,
         info: Info[GraphQLContext, 'AssayMutations'],
-    ) -> Annotated['GraphQLComment', strawberry.lazy('api.graphql.schema')]:
+    ) -> Annotated[GraphQLComment, strawberry.lazy('api.graphql.schema')]:
         """Add a comment to a assay"""
-        # Import needed here to avoid circular import
-        from api.graphql.schema import GraphQLComment
-
         connection = info.context['connection']
         cl = CommentLayer(connection)
         result = await cl.add_comment_to_entity(
@@ -51,10 +47,8 @@ class AssayMutations:
     @strawberry.mutation
     async def create_assay(
         self, assay: AssayUpsertInput, info: Info
-    ) -> Annotated['GraphQLAssay', strawberry.lazy('api.graphql.schema')]:
+    ) -> Annotated[GraphQLAssay, strawberry.lazy('api.graphql.schema')]:
         """Create new assay, attached to a sample"""
-        from api.graphql.schema import GraphQLAssay
-
         connection = info.context['connection']
         alayer = AssayLayer(connection)
         upserted = await alayer.upsert_assay(
@@ -70,8 +64,6 @@ class AssayMutations:
         info: Info,
     ) -> Annotated['GraphQLAssay', strawberry.lazy('api.graphql.schema')]:
         """Update assay for ID"""
-        from api.graphql.schema import GraphQLAssay
-
         if not assay.id:
             raise ValueError('Assay must have an ID to update')
         connection = info.context['connection']
