+++ api/graphql/mutations/family.py
@@ -1,18 +1,15 @@
-# pylint: disable=redefined-builtin, import-outside-toplevel
-
-from typing import TYPE_CHECKING, Annotated
+from typing import Annotated
 
 import strawberry
-from strawberry.types import Info
 from strawberry.scalars import JSON
+from strawberry.types import Info
 
 from api.graphql.loaders import GraphQLContext
+from api.graphql.types.comments import GraphQLComment
+from api.graphql.types.family import GraphQLFamily
 from db.python.layers.comment import CommentLayer
 from db.python.layers.family import FamilyLayer
 from models.models.comment import CommentEntityType
-
-if TYPE_CHECKING:
-    from api.graphql.schema import GraphQLComment, GraphQLFamily
 
 
 @strawberry.input
@@ -33,17 +30,14 @@ class FamilyMutations:
     async def add_comment(
         self,
         content: str,
-        id: int,
+        family_id: int,
         info: Info[GraphQLContext, 'FamilyMutations'],
-    ) -> Annotated['GraphQLComment', strawberry.lazy('api.graphql.schema')]:
+    ) -> Annotated[GraphQLComment, strawberry.lazy('api.graphql.schema')]:
         """Add a comment to a family"""
-        # Import needed here to avoid circular import
-        from api.graphql.schema import GraphQLComment
-
         connection = info.context['connection']
         cl = CommentLayer(connection)
         result = await cl.add_comment_to_entity(
-            entity=CommentEntityType.family, entity_id=id, content=content
+            entity=CommentEntityType.family, entity_id=family_id, content=content
         )
         return GraphQLComment.from_internal(result)
 
@@ -52,10 +46,8 @@ class FamilyMutations:
         self,
         family: FamilyUpdateInput,
         info: Info,
-    ) -> Annotated['GraphQLFamily', strawberry.lazy('api.graphql.schema')]:
+    ) -> Annotated[GraphQLFamily, strawberry.lazy('api.graphql.schema')]:
         """Update information for a single family"""
-        from api.graphql.schema import GraphQLFamily
-
         connection = info.context['connection']
         flayer = FamilyLayer(connection)
         await flayer.update_family(
