+++ api/graphql/mutations/sample.py
@@ -1,6 +1,4 @@
-# pylint: disable=redefined-builtin, import-outside-toplevel
-
-from typing import TYPE_CHECKING, Annotated
+from typing import Annotated
 
 import strawberry
 from strawberry.types import Info
@@ -8,19 +6,17 @@ from strawberry.types import Info
 from api.graphql.loaders import GraphQLContext
 from api.graphql.mutations.assay import AssayUpsertInput
 from api.graphql.mutations.sequencing_group import SequencingGroupUpsertInput
+from api.graphql.types.comments import GraphQLComment
+from api.graphql.types.sample import GraphQLSample
 from db.python.connect import Connection
 from db.python.layers.comment import CommentLayer
 from db.python.layers.sample import SampleLayer
 from models.models.comment import CommentEntityType
+from models.models.project import FullWriteAccessRoles
 from models.models.sample import SampleUpsert
 from models.utils.sample_id_format import (  # Sample,
     sample_id_transform_to_raw,
 )
-from models.models.project import FullWriteAccessRoles
-
-
-if TYPE_CHECKING:
-    from api.graphql.schema import GraphQLComment, GraphQLSample
 
 
 @strawberry.input  # type: ignore [misc]
@@ -46,17 +42,15 @@ class SampleMutations:
     async def add_comment(
         self,
         content: str,
-        id: str,
+        sample_id: str,
         info: Info[GraphQLContext, 'SampleMutations'],
-    ) -> Annotated['GraphQLComment', strawberry.lazy('api.graphql.schema')]:
+    ) -> Annotated[GraphQLComment, strawberry.lazy('api.graphql.schema')]:
         """Add a comment to a sample"""
-        from api.graphql.schema import GraphQLComment
-
         connection = info.context['connection']
         cl = CommentLayer(connection)
         result = await cl.add_comment_to_entity(
             entity=CommentEntityType.sample,
-            entity_id=sample_id_transform_to_raw(id),
+            entity_id=sample_id_transform_to_raw(sample_id),
             content=content,
         )
         return GraphQLComment.from_internal(result)
@@ -69,10 +63,8 @@ class SampleMutations:
         project: str,
         sample: SampleUpsertInput,
         info: Info[GraphQLContext, 'SampleMutations'],
-    ) -> Annotated['GraphQLSample', strawberry.lazy('api.graphql.schema')]:
+    ) -> Annotated[GraphQLSample, strawberry.lazy('api.graphql.schema')]:
         """Creates a new sample, and returns the internal sample ID"""
-        from api.graphql.schema import GraphQLSample
-
         connection: Connection = info.context['connection']
         (target_project,) = connection.get_and_check_access_to_projects_for_names(
             [project], FullWriteAccessRoles
@@ -93,13 +85,11 @@ class SampleMutations:
         project: str,
         samples: list[SampleUpsertInput],
         info: Info[GraphQLContext, 'SampleMutations'],
-    ) -> list[Annotated['GraphQLSample', strawberry.lazy('api.graphql.schema')]] | None:
+    ) -> list[Annotated[GraphQLSample, strawberry.lazy('api.graphql.schema')]] | None:
         """
         Upserts a list of samples with sequencing-groups,
         and returns the list of internal sample IDs
         """
-        from api.graphql.schema import GraphQLSample
-
         connection: Connection = info.context['connection']
         (target_project,) = connection.get_and_check_access_to_projects_for_names(
             [project], FullWriteAccessRoles
@@ -128,10 +118,8 @@ class SampleMutations:
         self,
         sample: SampleUpsertInput,
         info: Info[GraphQLContext, 'SampleMutations'],
-    ) -> Annotated['GraphQLSample', strawberry.lazy('api.graphql.schema')]:
+    ) -> Annotated[GraphQLSample, strawberry.lazy('api.graphql.schema')]:
         """Update sample with id"""
-        from api.graphql.schema import GraphQLSample
-
         connection = info.context['connection']
         slayer = SampleLayer(connection)
         upserted = await slayer.upsert_sample(
